backend/requirements.txt
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT
flask-marshmallow
marshmallow-sqlalchemy

